# [3604] Make node overlap resolution faster during "Arrange All"