- **[3604] Update CHANGELOG**
- **[3604] Only compute if nodes intersect**
- **[3604] Remove useless test**
- **[3604] Filter out non-siblings earlier**
- **[3604] Avoid creating intermediate rectangles**

On the following diagram (a simple diagram with recursive nodes for every `papaya::NamedElement`, created on the "Package java.lang" element in the "Sirius Web" Papaya example), I go from 60s to a little less than 20s.

![image](https://github.com/eclipse-sirius/sirius-web/assets/10608/4ec65183-e978-446a-b3ce-cae4d32047e9)

Note that the first 6s or 7s seem to be spent in ELK? At least during this time I see the loading indicator added in #3574 spinning for these first seconds, and then it stops moving while my CPU is working for the rest of the duration.

The diagram is very specific and does not contain any actual overlap, so it's very possible my changes breaks some stuff. Take with a grain of salt for the moment.