@pcdavid 2024-06-17
After this first batch, a profiling in Chromium shows this:

![Capture d’écran du 2024-06-17 17-40-19](https://github.com/eclipse-sirius/sirius-web/assets/10608/34c85296-9b9b-41c0-8eec-93eb32bf8f1e)

* First (in red), about 7s waiting for the backend for the result of the `mutation layoutDiagram()` operation.
* Then (in blue), about 13s trying to detect and resolve overlaps (which do not exist in this case).

I'm wondering if the backend could pre-compute some of the required information? Even just a flag indicating if there *are* any overlap so that the front could simply avoid all of this in the (common?) case where there are none.